20160203客服系统BUG-微信支付帐号查询报错 http://tapd.oa.com/tenpay_kf/prong/stories/view/1010068761057420603

+++ src/CFT.CSOMS.BLL/TradeModule/TradeService.cs
@@ -48,7 +48,7 @@ namespace CFT.CSOMS.BLL.TradeModule
             }
         }
        
-        public DataSet QueryWxBuyOrderByUid(int uid, DateTime startTime, DateTime endTime)
+        public DataSet QueryWxBuyOrderByUid(Int64 uid, DateTime startTime, DateTime endTime)
         {
             return (new TradeData()).QueryWxBuyOrderByUid(uid, startTime, endTime);
         }
